requirements.txt
langgraph==0.6.7
langchain==0.3.27
langchain-openai==0.3.33
gradio==5.46.0
python-dotenv==1.1.1